requirements.txt
aiohttp==3.9.0
groq==0.9.0
python-dotenv==1.0.0
